+++ CHANGES.rst
@@ -11,6 +11,8 @@ Breaking changes:
 
 - Switch from optparse to argparse to work with zope.testrunner >= 4.9.0
   [pbauer]
+- Drop python 2.6 dependency.
+  [gforcada]
 
 New features:
 
@@ -18,8 +20,8 @@ New features:
 
 Bug fixes:
 
-- *add item here*
-
+- Clean up the code.
+  [gforcada]
 
 1.3.4 (2017-02-02)
 ------------------